create seperate usehook file to useFamilyMembersTable
